Add shared cloth textures and update Murlan table felt
